Backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "divine-whispers-backend"
version = "1.0.0"
description = "Fortune-telling application backend with ChromaDB integration"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Divine Whispers", email = "contact@divinewhispers.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # FastAPI and ASGI server
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    
    # Database
    "sqlalchemy==2.0.23",
    "asyncpg==0.29.0",
    "alembic==1.13.0",
    
    # Authentication and Security
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    
    # Configuration
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    
    # HTTP client
    "httpx==0.25.2",
    "aiohttp==3.9.1",
    
    # Validation and serialization
    "email-validator==2.1.0",
    "phonenumbers==8.13.26",
    
    # Background tasks and async (Note: Removed Redis/Celery as we don't use them)
    "asyncio-mqtt==0.11.1",  # For async background tasks
    
    # ChromaDB integration
    "chromadb==0.4.18",
    "sentence-transformers==2.2.2",
    
    # LLM integration
    "openai==1.3.7",
    "tiktoken==0.5.2",
    
    # Utilities
    "python-dateutil==2.8.2",
    "pytz==2023.3",
    "cachetools==5.3.2",  # For in-memory caching
    
    # Logging and monitoring
    "structlog==23.2.0",
    "prometheus-client==0.19.0",
    
    # File handling
    "python-magic==0.4.27",
    "pillow==10.1.0",
    
    # Rate limiting
    "slowapi==0.1.9",
]

[project.optional-dependencies]
dev = [
    # Development and testing
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1", 
    "pytest-mock==3.12.0",
    "faker==20.1.0",
    
    # Code quality
    "black==23.12.0",
    "isort==5.13.2",
    "flake8==6.1.0",
    "mypy==1.7.1",
    
    # Development tools
    "pre-commit==3.6.0",
]

production = [
    # Production monitoring
    "sentry-sdk[fastapi]==1.38.0",
    "gunicorn==21.2.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/divine-whispers"
Documentation = "https://docs.divinewhispers.com"
Repository = "https://github.com/yourusername/divine-whispers"
Issues = "https://github.com/yourusername/divine-whispers/issues"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "sentence_transformers.*",
    "tiktoken.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"